requirements.txt
Flask==3.0.3
mysql_connector_repackaged==0.3.1
python-dotenv==1.0.1
Werkzeug==3.0.2
